control-system/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
asyncpg==0.30.0
billiard==4.2.1
black @ git+https://github.com/psf/black@7987951e246b4e76cc5225b2ccd89b5519a25ac8
celery==5.5.2
certifi==2025.4.26
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dill==0.4.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
isort==6.0.1
Jinja2==3.1.6
joblib==1.5.1
kombu==5.5.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.2.6
nvidia-nccl-cu12==2.27.3
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.0
pathspec==0.12.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
pylint==3.3.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.2
redis==5.3.0
rich==14.0.0
rich-toolkit==0.14.6
scikit-learn==1.6.1
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
xgboost==3.0.2
